+++ src/main/java/org/cthing/versionparser/internal/annotations/PackageNonnullByDefault.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cthing.versionparser.annotations;
+package org.cthing.versionparser.internal.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;