+++ src/main/java/org/cthing/versionparser/internal/annotations/UtilityClass.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.cthing.versionparser.annotations;
+package org.cthing.versionparser.internal.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;