+++ src/main/java/org/cthing/versionparser/maven/MvnVersionScheme.java
@@ -43,7 +43,7 @@ import org.cthing.versionparser.Version;
 import org.cthing.versionparser.VersionConstraint;
 import org.cthing.versionparser.VersionParsingException;
 import org.cthing.versionparser.VersionRange;
-import org.cthing.versionparser.annotations.UtilityClass;
+import org.cthing.versionparser.internal.annotations.UtilityClass;
 
 
 /**